Fix comment, move some files back
