+++ packages/roosterjs-editor-plugins/lib/index.ts
@@ -4,3 +4,11 @@ export { default as HyperLink } from './HyperLink/HyperLink';
 export { default as ContentEdit, TabIndent } from './ContentEdit/ContentEdit';
 export { default as PasteManager, PasteOption } from './PasteManager/PasteManager';
 export { default as Watermark } from './Watermark/Watermark';
+export {
+    default as ClipBoardData,
+    ImageData,
+    LocalFileImageData,
+    Base64ImageData,
+} from './PasteManager/ClipboardData';
+export { getPastedElementWithId } from './PasteManager/processImages';
+export { default as buildClipBoardData } from './PasteManager/buildClipBoardData';
